third_party/blink/web_tests/platform/win/external/wpt/css/css-fonts/variations/at-font-face-font-matching-expected.txt

This is a testharness.js-based test.
Found 11 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Descriptor matching priority: Style has higher priority than weight
  assert_equals: Unexpected font on test element expected 442 but got 487
[FAIL] Matching font-style: 'normal' should prefer 'normal' over 'oblique 0deg'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'italic' should prefer 'normal' over 'oblique 0deg'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 20deg' should prefer 'oblique 10deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 21deg' should prefer 'oblique 10deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 10deg' should prefer 'oblique 40deg 50deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 10deg' should prefer 'italic' over 'oblique 0deg'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 0deg' should prefer 'oblique 40deg 50deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique -10deg' should prefer 'italic' over 'oblique 0deg 10deg'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique -20deg' should prefer 'italic' over 'oblique 0deg'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique -21deg' should prefer 'italic' over 'oblique 0deg'
  assert_equals: Unexpected font on test element expected 487 but got 532
Harness: the test ran to completion.
